Extract database code back into package
