--- magic_packet/cli/common_voice/database/__init__.py
@@ -1,4 +1,0 @@
-from .database_manager import DatabaseManager
-from .records import Clips, Words, sql_join
-
-__all__ = ["Clips", "DatabaseManager", "sql_join", "Words"]
